Match C# parser

+++ src/java/org/deltava/beans/simbrief/SimBriefParser.java
@@ -18,7 +18,7 @@ import org.deltava.util.system.SystemData;
 /**
  * A parser for SimBrief XML dispatch packages.
  * @author Luke
- * @version 12.3
+ * @version 12.4
  * @since 10.3
  */
 
@@ -100,7 +100,7 @@ public class SimBriefParser {
 		
 		// Load payload
 		sb.setPax(StringUtils.parse(XMLUtils.getChildText(re, "weights", "pax_count_actual"), 0));
-		sb.setBagWeight(StringUtils.parse(XMLUtils.getChildText(re, "weights", "bag_weight"), 55));
+		sb.setBagWeight((int) StringUtils.parse(XMLUtils.getChildText(re, "weights", "bag_weight"), 55d));
 		sb.setBaggageWeight(StringUtils.parse(XMLUtils.getChildText(re, "weights", "bag_count_actual"), 0) * sb.getBagWeight());
 		sb.setCargoWeight(StringUtils.parse(XMLUtils.getChildText(re, "weights", "freight_added"), 0));
